Crate integration_tests

Integration testing

Modules

integration_test_circuits

Common code for integration tests of circuits.

Structs

CompiledContract

Solc-compiled contract output

GETH0_URL

URL of the integration test geth0 instance, which contains blocks for which proofs will be generated.

GenDataOutput

Output information of the blockchain data generated by gen_blockchain_data.

Constants

CHAIN_ID

Geth dev chain ID

CONTRACTS

List of contracts as (ContractName, ContractSolidityFile)

CONTRACTS_PATH

Path to the test contracts

GENDATA_OUTPUT_PATH

Path to gen_blockchain_data output file

WARN

Solidity compilation warnings to ignore (by error code) 2018: Warning - “Function state mutability can be restricted to pure” 5667: Warning - “Unused function parameter. Remove or comment out the variable name to silence this warning.” For smart contracts that are optimized for worst case block generation, we want to allow contracts that do not interfere with state, without setting state mutability to view. otherwise compiler optimizations will not allow recursive execution of targeted opcodes

Functions

get_chain_id

Get the chain id by querying the geth client.

get_client

Get the integration test [GethClient]

get_provider

Get the integration test [Provider]

get_wallet

Get a wallet by index

log_init

Initialize log